+++ webpack.prod.config.js
@@ -12,6 +12,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 const CopyWebpackPlugin = require('copy-webpack-plugin')
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin')
+const ScriptExtHtmlWebpackPlugin = require('script-ext-html-webpack-plugin')
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
 
 let publicPath = `https://static.ws.126.net/163/f2e/${pkg.channel}/${pkg.name}/`
@@ -33,10 +34,15 @@ module.exports = merge(baseWebpackConfig, {
         template: "./src/index.html",
         filename: "./index.html"
     }),
+    new ScriptExtHtmlWebpackPlugin({
+      inline: /manifest\..*\.js$/,
+      async: /main\..*\.js$/,
+    }),
     new webpack.NamedModulesPlugin(),
     new webpack.DefinePlugin({
         'process.env.BASE_URL': JSON.stringify(publicPath),
         'process.env.ANT_PROJECT_ID': JSON.stringify(pkg.projectId),
+        'process.env.PROJECT_CHANNEL': JSON.stringify(pkg.channel),
         'process.env.PROJECT_NAME': JSON.stringify(pkg.name)
     }),
     new CopyWebpackPlugin([